// src/app/model/slack-user.ts
export class SlackUser {
    id: any;
    name: any;
    username: any;
    password: any;
    image: any;
    user_type: any;
    remember_token: any;
    slack_user_id: any;
    token: any;
    bot_acess_token: any;
    web_hook_url: any;
    created_at: any;
    updated_at: any;
    workspace_id: any;
    user_lists: any;
    mytheme: any; 
}